create header and reoganize code to avoid having everything into one class
